Files created + getComputerChoice
managed to create the files after having looked back at how to do it exactly
been quite some time seeing as i went on holiday for 2 weeks. Also managed to
create the first necessary function to get the computers choice. Did need to use
some google which feels like cheating but on the other hand i havent had many of
the ways to do this.